+++ src/main/java/fu/swp/dorm_mnm/repository/RoomRepository.java
@@ -1,10 +1,12 @@
 package fu.swp.dorm_mnm.repository;
 
-import fu.swp.dorm_mnm.model.Room;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import fu.swp.dorm_mnm.model.Room;
+
+import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface RoomRepository extends JpaRepository<Room, Integer> {
+public interface RoomRepository extends JpaRepository<Room, Long>{
+    
 }
