+++ src/main/java/fu/swp/dorm_mnm/repository/RoomTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoomTypeRepository extends JpaRepository<RoomType, Integer> {
+public interface RoomTypeRepository extends JpaRepository<RoomType, Long> {
 }
